move query input component to its own folder
