+++ src/app/views/query-runner/query-input/index.ts
@@ -1,0 +1,3 @@
+import QueryInput from './QueryInput';
+
+export { QueryInput };
